Consistency of casing in unit tests
