+++ src/lib/appContainer.ts
@@ -83,7 +83,7 @@ export const getAppContainer = (overrides?: IAppContainer): IAppContainer => {
     get accountHolderApiGateway() {
       return new BeaconsApiAccountHolderGateway(process.env.API_URL);
     },
-    get userSessionGateway() {
+    get NextAuthUserSessionGateway() {
       return new NextAuthUserSessionGateway();
     },
     get draftRegistrationGateway() {
