+++ test/deepFreeze.ts
