+++ test/fixtures/accountHolder.fixture.ts
@@ -1,5 +1,5 @@
 import { AccountHolder } from "../../src/entities/AccountHolder";
-import { deepFreeze } from "../utils/deepFreeze";
+import { deepFreeze } from "../deepFreeze";
 
 export const accountHolderFixture: AccountHolder = deepFreeze({
   id: "cb2e9fd2-45bb-4865-a04c-add5bb7c34a7",
