+++ test/fixtures/emergencyContacts.fixture.ts
@@ -1,5 +1,5 @@
 import { EmergencyContact } from "../../src/entities/EmergencyContact";
-import { deepFreeze } from "../utils/deepFreeze";
+import { deepFreeze } from "../deepFreeze";
 
 export const emergencyContactsFixture: EmergencyContact[] = deepFreeze([
   {
