+++ test/fixtures/manyBeaconsApiResponse.fixture.ts
@@ -1,5 +1,5 @@
 import { IBeaconListResponse } from "../../src/gateways/mappers/IBeaconListResponse";
-import { deepFreeze } from "../utils/deepFreeze";
+import { deepFreeze } from "../deepFreeze";
 
 const getUseResponseJson = (mainUse: boolean) =>
   deepFreeze({
