+++ test/gateways/BeaconsApiResponseMapper.test.ts
