+++ test/lib/registration/DeprecatedRegistration.ts
